minor changes
git-svn-id: svn://141.76.65.164/kp-salespoint@110 eb53f5e8-9ec5-4bc0-8d13-06c033283886

+++ SalespointFramework/src/org/salespointframework/core/product/AbstractProductInstance.java
@@ -17,10 +17,10 @@ public abstract class AbstractProductInstance<T extends AbstractProductType> imp
 	
 	@Id
 	@GeneratedValue
-	private int serialNumber;
+	private long serialNumber;
 	
 	// TODO richtig?
-	//@ManyToOne
+	//@OneToOne
 	private T productType;
 	
 	private Money price;
@@ -55,7 +55,7 @@ public abstract class AbstractProductInstance<T extends AbstractProductType> imp
 	}
 
 	@Override
-	public int getSerialNumber() {
+	public long getSerialNumber() {
 		return serialNumber;
 	}
